refactor : update to prize statistics access
