+++ src/main/java/lotto/domain/prize/WinningReport.java
@@ -1,6 +1,8 @@
-package lotto.domain;
+package lotto.domain.prize;
 
 import lotto.constant.Prize;
+import lotto.domain.number.Lotto;
+import lotto.domain.sales.LottoBundle;
 
 public class WinningReport {
 
@@ -9,7 +11,7 @@ public class WinningReport {
     private final Lotto winningLotto;
     private final WinningPrize winningPrize;
 
-    public WinningReport(int salesAmount, LottoBundle lottoBundle, Lotto winningLotto) {
+    private WinningReport(int salesAmount, LottoBundle lottoBundle, Lotto winningLotto) {
         this.salesAmount = salesAmount;
         this.lottoBundle = lottoBundle;
         this.winningLotto = winningLotto;
@@ -17,10 +19,13 @@ public class WinningReport {
     }
 
     public static WinningReport of(int salesAmount, LottoBundle lottoBundle, Lotto winningLotto) {
-        return new WinningReport(salesAmount, lottoBundle, winningLotto);
+        WinningReport report = new WinningReport(salesAmount, lottoBundle, winningLotto);
+        report.analyze();
+
+        return report;
     }
 
-    public void analyze() {
+    private void analyze() {
         for (int i = 0; i < lottoBundle.count(); i++) {
             int match = winningLotto.match(lottoBundle.lotto(i));
             Prize prize = Prize.getPrize(match);
@@ -32,6 +37,10 @@ public class WinningReport {
         if (prize != null) {
             winningPrize.add(prize);
         }
+    }
+
+    public int winAmount(Prize prize) {
+        return winningPrize.winAmount(prize);
     }
 
     public int first() {
